pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "casictool"
version = "0.1.0"
description = "CASIC GPS receiver configuration tool"
requires-python = ">=3.9"
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-xdist>=3.0.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "types-pyserial",
]

[project.scripts]
casictool = "casictool:main"

[tool.setuptools]
py-modules = ["casictool", "casic"]

[tool.mypy]
strict = true

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
